tests and reverting logic
reverting the logic since it really doesnt make any sense to scale to a job when a simple function can do the heavy lifting. if we ever get approached by a website with a desist of the scraping then we can implement it, but in the meantime its a no go. Also added some tests